Add files via upload
Geolocation Data
Geospatial information is data that is referenced by spatial or geographic coordinates. The data that we will be working with in this lesson is vector data - features that are represented by points, lines, and polygons.

Points are defined by a pair of (x,y) coordinates. They usually represent locations, place names, and other objects on the ground.
Lines are the connection between two points. They can have properties such as length, direction, flow, etc.
Polygons are a series of lines connected together to form a shape. They can have properties such as area, perimeters, and centroids.
In this notebook, you will need to install the geopandas and geoPy libraries.